+++ modules/data.land/R/find.land.R
@@ -11,7 +11,14 @@
 ##' @export
 ##' @author David LeBauer
 find.land <- function(lat, lon, plot = FALSE) {
-  data("wrld_simpl",package="maptools",envir = environment())
+  if (!requireNamespace("maptools", quietly = TRUE)) {
+    PEcAn.logger::logger.error(
+      "PEcAn.data.land::find.land needs package `maptools`.",
+      "Please install it and try again.")
+  }
+  # If maptools used lazy data, this could just be `maptools::wrld_simpl`...
+  wrld_simpl <- NULL
+  utils::data("wrld_simpl",package="maptools",envir = environment())
   
   ## Create a SpatialPoints object
   points <- expand.grid(lon, lat)
@@ -19,7 +26,7 @@ find.land <- function(lat, lon, plot = FALSE) {
   pts <- sp::SpatialPoints(points, proj4string = sp::CRS(sp::proj4string(wrld_simpl)))
   
   ## Find which points fall over land
-  landmask <- cbind(points, data.frame(land = !is.na(over(pts, wrld_simpl)$FIPS)))
+  landmask <- cbind(points, data.frame(land = !is.na(sp::over(pts, wrld_simpl)$FIPS)))
   if (plot) {
     plot(wrld_simpl)
     landmask[, points(lon, lat, col = 1 + land, pch = 16)]